fix(debugger): readUint16() 返回 int, 赋给 short 需强转
DwarfSymbolResolver.java 两处:
  line 254: short version = r.readUint16();
  line 349: short lver = lr.readUint16();

readUint16() 返回 int (0..65535), Java 不允许隐式 int→short
转换 (可能丢失精度), 导致 compileDebugJavaWithJavac 失败。
加 (short) 强转。

+++ debugger/Breakpoint-debugger/ide-debugger/src/main/java/com/zerostudio/debugger/symbol/DwarfSymbolResolver.java
@@ -251,7 +251,7 @@ public final class DwarfSymbolResolver implements SourceNameMapper.SymbolResolve
                         if (is64Unit) { totalLen = r.readUint64(); } else { totalLen = unitLen; }
                         if (totalLen == 0) break;
                         long unitEnd = r.position() + totalLen - (is64Unit ? 12 : 4);
-                        short version = r.readUint16();
+                        short version = (short) r.readUint16();
                         r.readUint32(); // debug_info offset
                         if (is64) r.readUint64(); // 8-byte address
                         r.readUint8(); // address size
@@ -346,7 +346,7 @@ public final class DwarfSymbolResolver implements SourceNameMapper.SymbolResolve
                         long total = is64L ? lr.readUint64() : unitLen;
                         if (total == 0) break;
                         long unitEnd = lr.position() + total - (is64L ? 12 : 4);
-                        short lver = lr.readUint16();
+                        short lver = (short) lr.readUint16();
                         if (lver < 2 || lver > 5) { lr.seek(unitEnd); continue; }
                         int hdrLen = is64L ? (int) lr.readUint64() : lr.readUint32();
                         long afterHdr = lr.position() + hdrLen;
